Parse a FQPN at any position in the package list
This patch is a step forward in correct handling multiple packages. With this patch if a FullQualifiedPackageName get passed in the
package list it get parsed no matter the position (before it only happened for the first one).
Also, the first package keep its importance of potential determining the ami-name or the Appversion if correctly parsed