Improve the unicity of the bake key also taking into account multiple accounts
This PR address the following issue: Given two bakes with the same ami-name but different installed package(s), the bakeKey should be different, otherwise a rebake is required to bake it again (also the ami-suffix is not part of the bakeKey).
This is very evident when baking from the Bake stage of a pipeline. The ami-name is constant at this point, while the installed packages version is different for each build