feat(bake/windows): Allow Chocolatey version control via attributes
The current Chocolatey scripts always use the latest version, meaning
that bugs in new releases of Chocolatey may affect Windows bakes. This
pull allows a "chocolateyVersion" attribute to be applied to a bake step
to manually control the version installed by the scripts.

Default behavior is unchanged, it will install the latest version.