fix(packer): Fix spot pricing flags
Packer 1.3 now enforces that aws_spot_price_auto_product can only
be passed if aws_spot_price is set to "auto". The default settings
that ship with Spinnaker set aws_spot_price to "0" and
aws_spot_price_auto_product non-null. Gate passing through the
product value so that it is only passed through if the spot price
is "auto".