fix(artifacts): Fix NPE on empty artifact
If an artifact is empty, we throw an NPE when downloading it
rather than just pass an empty artifact along to the caller.

In general it's not overly useful to have an empty artifact, but
in some cases (ex: a values.yaml file for a Helm bake) it's reasonable
to have no content.