fix(bake): Improve error message
When supplying an artifact without a reference, we currently
throw an NPE that is hard to debug. Replace this with a better
error message.

(At some point we could consider if we even need a reference at
all, or if we could fall back to something else, but for now this
will at least make the restriction clearer.)