refactor(core): Convert package name logic to java
When trying to upgrade to Spring Boot 2 (and along with it
groovy 2.5) I learned that Groovy changed the .pop() method
from removing the last element to removing the first element.

This broke some of the package name logic, which used the .pop()
method. Sidestep this entirely by converting it to java, as it's
well-tested enough this can be done with reasonable confidence.