test(core): Fix tests that break with groovy upgrade
A few of the rosco tests fail when we upgrade groovy because
groovy no longer considers objects equal they are identical
except that a field is null in one vs empty array in the other.

That was masking the fact that we weren't stubbing out the
getMaskedPackerParameters function on our mock CloudProviderBakeHandler.
Add in that stubbing, and also change the calling code to use the
actual accessor rather than the groovy-generated synthetic one.

There's probably a way to factor out some of the common logic in the
tests so changes like this don't need to touch as many lines, but
for now just making the required change in each test.