feat(aws/ami_regions): add support for copies of amis
With packer one can copy amis from one bake to different regions. This
way one can have one bake and use the same artifact in different
regions.

In the issue spinnaker/spinnaker#4641 this feature is requested.

This pull request keeps backwards compatability. It starts collecting
all amis that are produced in a bake and stores them in a list. It adds
a field `regions` to the api that will allow spinnaker/orca to pick up
that we have amis for different regions. In the bake endpoint in rosco
we start using region to filter through the various amis in artifacts if
it is present.

There needs to be done changes to orca as well in order to use this.