fix(helm): Fix baking of helm artifacts (#442)
We currently throw an exception if a helm artifact doesn't have a
reference. We're only using that reference to generate a file name
to use when downloading the artifact; helm artifacts don't actually
need a reference.

Fix this by just falling back to the name-version if there is no
reference defined.