refactor(manifest): Encapsulate paths in BakeManifestEnvironment
Manifest bakes (helm, customize) create a BakeManifestEnviornment
that encapsulates a staging directory. Instead of exposing the
staging directory to consumers and requiring them to do the path
operations, hide the actual staging path from consumers and instead
expose operations to resovle paths relative to that path.

A nice advantage of this is that we can pull the logic to ensure
consumers don't break out of the staging path into the class.