refactor(manifest): Avoid passing around a byte[]
We're returning the result of a manifest bake as a byte[] because
we'll eventually need to get the bytes to encode the result as
base64. Instead of having doBake() return the byte[] directly, have
it return a String and only convert to bytes right before encoding as
base64.

This reduces the number of functions that need to accept/return a
byte[]. It also removes the extra conversion that needed to happen in
removeTestsDirectoryTemplates.