refactor(manifest): Don't inherit artifact downloader
We have a base class TemplateUtils from which both the Helm* and
Kustomize* template utils inherit. The only functionality here is
an artifact download function; instead of having helm and kustomize
inherit from this base class, just call it ArtifactDownloader and
compose it into the other utils classes.

This also cleans up some of the mocking/stubbing in the utils classes
(as now they can just stub the result from downloadArtifact instead
of the clouddriver response) and also allows us to add some tests
on the new ArtifactDownloader.