chore(gcb build): remove a symlink
There was a symlink `rosco-web/config/packer -> halconfig/packer`. But
GCB doesn't play well with symlinks. It uploads it as an empty
directory instead. While the symlink is nice, it's not particularly
important, so I just removed it. I put the files in `halconfig/packer`
and reference them there in the build script and dockerfiles.