refactor(artifacts): Use toBuilder rather than copy all fields
Another change to artifacts is that we will no longer allow access
to the full metadata map via .getMetadata and will instead require
callers to ask for they key they want via getMetadata(String key).

In the one call here, we were only using getMetadata to copy the
data over to a new artifact; this can be replaced by using the
new toBuilder function which copies over all fields. We can then
set the one field that we weren't explicitly setting to the
value of the original artifact.