fix(manifests): catch SpinnakerException how that rosco uses SpinnakerRetrofitErrorHandler when talking to clouddriver
https://github.com/spinnaker/rosco/pull/750 changed rosco to use
SpinnakerRetrofitErrorHandler when talking to clouddriver.  As a consequence, exceptions
changed from RetrofitError to SpinnakerException, so update the code that looked for
RetrofitError, to restore the previous wrapping and error messages to rosco's callers.

While we're at it, change ArtifactDownloaderImpl.downloadArtifactToFile to throw a
SpinnakerException when it catches one since that seems more informative than IOException,
especially since downloadArtifactToFile throws IOExceptions for other reasons.