chore(ci): GHA - tidy 'spinnaker-' prefix and '-latest' usage
`spinnaker` is already in the container repository name so the prefix
`spinnaker-` doesn't add any meaning to our container tags so we can
remove it simplifying the tag.

When we are adding a git commit SHA and DATE TIME the string "-latest"
doesn't add any meaning so we can remove it, simplifying the tag.