feat(aws): Upgrade packer and add SSM support (Amazon)
This PR upgrades Packer to version v1.10.1. All the previously default providers were moved to plugins in v1.10 [1], so we also need to manually install plugins. I've opted to install Amazon, Azure, Docker and Google plugins by default, but the list can be overridden by Docker build args (or an environment variable in the case of the deb package).

This PR also installs AWS CLI 2 and the Session Manager Plugin in the Ubuntu images and in the deb package. **I opted to not install it in the Alpine images**, as the session-manager-plugin is not officially supported, and you'd need to either copy the binary or compile it from source to get it to work. I'm very open to reconsider this.

After applying this PR, you can use SSM instead of SSH to connect to instances when baking using the `amazon-ebs` provider (or friends) [2].

[1]: https://www.hashicorp.com/blog/announcing-the-removal-of-bundled-plugins-in-hashicorp-packer
[2]: https://developer.hashicorp.com/packer/integrations/hashicorp/amazon/latest/components/builder/ebs#session-manager-plugin