fix(helm) : handle overrides with nested keys
When creating overrides YAML files with nested key structures, they are being represented as flattened rather than structured keys. This leads to discrepancies in the output of the helm template command when comparing the use of --set/--set-string versus --values with an overrides YAML. While --set/--set-string correctly renders keys with their nested structure, using --values with a YAML file resulted in the omission of these nested keys.  This fixes that, so nested keys work.

For example, if an override was specified as
```
overrides:
  image.tag: "foo"
```
before this PR, the value of image.tag wouldn't be "foo", it would be the default value in the helm chart.  With this PR, the value of image.tag is "foo".

A new HelmSetArgumentParser class inspired by Helm's Go code for parsing --set arguments handles complex nested structures, including lists and maps, and supports type inference for values.