fix(manifests): restore behavior that leaves float/double override values quoted
The reason to leave floats quoted is that it more closely matches helm's behavior.

Consider this helm chart:
```
apiVersion: v2
name: float-example
description: A Helm chart example that includes a float value
version: 0.1.0
appVersion: "1.0"
```
that has a values file internal to the chart (aka defaults) like:
```
threshold: 3.14
```
and
```
apiVersion: v1
kind: ConfigMap
metadata:
 name: float-example-config
data:
 # Using the float directly as a numeric value
 numericThreshold: {{ .Values.threshold }}

 # Converting the float to a string
 stringThreshold: "{{ .Values.threshold }}"

 # Alternatively, using Helm's `quote` function
 quotedThreshold: {{ quote .Values.threshold }}

 thresholdTypeSprig: "{{ typeOf .Values.threshold }}"
```

With both helm 3.13.1 and 3.16.2, look at the output of various commands:
```
$ helm template float-example ./float-example
---
apiVersion: v1
kind: ConfigMap
metadata:
  name: float-example-config
data:
  # Using the float directly as a numeric value
  numericThreshold: 3.14

  # Converting the float to a string
  stringThreshold: "3.14"

  # Alternatively, using Helm's `quote` function
  quotedThreshold: "3.14"

  thresholdTypeSprig: "float64"
```
So far so good: there's a float value in the chart, and it appears in the output as a float.

Now let's try overriding with an int value using --set:
```
$ helm template float-example ./float-example --set threshold=6
---
apiVersion: v1
kind: ConfigMap
metadata:
  name: float-example-config
data:
  # Using the float directly as a numeric value
  numericThreshold: 6

  # Converting the float to a string
  stringThreshold: "6"

  # Alternatively, using Helm's `quote` function
  quotedThreshold: "6"

  thresholdTypeSprig: "int64"
```
Still OK so far.  Now the clincher, let's try a decimal value using --set:
```
$ helm template float-example ./float-example --set threshold=6.28
---
apiVersion: v1
kind: ConfigMap
metadata:
  name: float-example-config
data:
  # Using the float directly as a numeric value
  numericThreshold: 6.28

  # Converting the float to a string
  stringThreshold: "6.28"

  # Alternatively, using Helm's `quote` function
  quotedThreshold: "6.28"

  thresholdTypeSprig: "string"
```
The type has changed to a string!

to complete the story, using --values:
```
$ cat external-values.yaml
threshold: 6.28
```
$ helm template float-example ./float-example --values external-values.yaml
---
apiVersion: v1
kind: ConfigMap
metadata:
  name: float-example-config
data:
  # Using the float directly as a numeric value
  numericThreshold: 6.28

  # Converting the float to a string
  stringThreshold: "6.28"

  # Alternatively, using Helm's `quote` function
  quotedThreshold: "6.28"

  thresholdTypeSprig: "float64"
```

To summarize:

- default (in-chart) value of float yields a float
- using --set threshold=6.28 yields a string
- using --values external-values.yaml with
```
    threshold: 6.28
```
yields a float

To tie this back to Spinnaker, an override in a helm bake manifest stage like:
```
threshold: 6.28
```
ends up as a `--set` option on the helm template command line before this PR, and so it's a string in the helm chart.  To preserve that behavior when the length of overrides is over the threshold, this PR has:
```
threshold: "6.28"
```
in the generated values file.

Removing the quotes in the generated values file seems cleaner, but is a change in the current behavior.  An alternate would be to add special logic to leave float overrides using `--set`.  That's left for a future PR.